% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super_utils.R
\name{check_package}
\alias{check_package}
\title{Internal function}
\usage{
check_package(pkg)
}
\arguments{
\item{pkg}{should be a string containing package name}
}
\value{
null
}
\description{
Used to check the package is installed
}
\keyword{internal}
