% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super_utils.R
\name{Counter}
\alias{Counter}
\title{Calculate count of values in a list}
\usage{
Counter(data, sort = TRUE, decreasing = FALSE)
}
\arguments{
\item{data}{should be a vector or list of input values}

\item{sort}{a logical value, to sort the result or not}

\item{decreasing}{a logical value, the order of sorting to be followed}
}
\value{
list of values in a sorted order
}
\description{
Handy to calculate count of values in a sorted order
}
\examples{
d <- list(c('i','am','bad'),c('you','are','also','bad'))
counts <- Counter(d, sort=TRUE, decreasing=TRUE)
}
