% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFolds.R
\name{createFolds}
\alias{createFolds}
\title{Internal function}
\usage{
createFolds(y, k = 10, list = TRUE, returnTrain = FALSE)
}
\arguments{
\item{y}{target variable}

\item{k}{number of folds}
}
\value{
null
}
\description{
Taken from caret package, used to create folds
}
\keyword{internal}
