% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dot}
\alias{dot}
\title{Dot product similarity in vectors}
\usage{
dot(a, b, norm = TRUE)
}
\arguments{
\item{a}{numeric vector}

\item{b}{numeric vector}

\item{norm}{logical, compute normalised dot product, default=True}
}
\value{
numeric vector containing sdot product score
}
\description{
Computes the dot product between two given vectors.
}
\examples{
a <- runif(5)
b <- runif(5)
s <- dot(a, b)

}
