% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dotmat}
\alias{dotmat}
\title{Dot product similarity between a vector and matrix}
\usage{
dotmat(a, b, norm = TRUE)
}
\arguments{
\item{a}{numeric vector}

\item{b}{numeric matrix}

\item{norm}{logical, compute normalised dot product, default=True}
}
\value{
numeric vector containing dot product scores
}
\description{
Computes the dot product between a vector and a given matrix.
The vector returned has a dot product similarity value for each row in the matrix.
}
