% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{normalise1d}
\alias{normalise1d}
\title{normalise1d}
\usage{
normalise1d(vec, pnorm = 2L)
}
\arguments{
\item{vec}{vector containing integers or numeric values.}

\item{pnorm}{integer, default: 2}
}
\value{
a vector containing normalised values
}
\description{
Normalises a 1 dimensional vector towards unit p norm. By default, p = 2 is used.
For a given vector, eg: c(1,2,3), norm value is calculated as `x / |x|` where `|x|` is calculated
as the square root of sum of square of values in the given vector.
}
\examples{
val <- c(1,10,5,3,8)
norm_val <- normalise1d(val)

}
