% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reg_train}
\alias{reg_train}
\title{reg_train}
\format{A data frame with 81 variables}
\source{
\url{https://www.kaggle.com/c/house-prices-advanced-regression-techniques/data}
}
\usage{
reg_train
}
\description{
Training Dataset used for regression examples. In this data set, we have to predict
the sale price of the houses.
}
\keyword{datasets}
