% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sort_index}
\alias{sort_index}
\title{sort_index}
\usage{
sort_index(vec, ascending = TRUE)
}
\arguments{
\item{vec}{numeric vector}

\item{ascending}{logical, order to return (ascending or descending), default = True}
}
\value{
numeric vector containing sorted indexes
}
\description{
For a given vector, return the indexes of the sorted array and
not the sorted array itself.
}
\examples{
v <- c(10,3,1,4)
j <- sort_index(v)

}
