% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{supersvd}
\alias{supersvd}
\title{Singular Value Decomposition (SVD)}
\usage{
supersvd(mat, n_components)
}
\arguments{
\item{mat}{matrix}

\item{n_components}{desired dimensionality of output data. Must be strictly less than the number of features.}
}
\value{
list, contains the following attributes:
        components: matrix
        explained_variance_ratio: Percentage of variance explained by each of the selected components.
        singular_values_array: The singular values corresponding to each of the selected components.
                               The singular values are equal to the 2-norms of the n_components variables in the lower-dimensional space.
}
\description{
Performs dimensionality reduction using svd by factorizing a given matrix
into a product of three matrics such that `A=UDV` where the columns of U and V are orthonormal
and the matrixDis diagonal with positive real entries.
}
\examples{
docs <- c("chimpanzees are found in jungle",
          "chimps are jungle animals",
          "Mercedes automobiles are best",
          "merc is made in germany",
          "chimps are intelligent animals")

sentence <- "automobiles are"
s <- bm_25(document=sentence, corpus=docs, top_n=2)

}
