% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super_utils.R
\name{testdata}
\alias{testdata}
\title{Internal function}
\usage{
testdata(X, y, model = NA)
}
\arguments{
\item{X}{should be a data frame or data.table}

\item{y}{should be a string specifying the dependent variable}
}
\value{
null
}
\description{
Used to check the input data format
}
\keyword{internal}
