% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cla_train}
\alias{cla_train}
\title{cla_train}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 891 rows and 12 columns.
}
\source{
\url{https://www.kaggle.com/c/titanic/data}
}
\usage{
cla_train
}
\description{
Training Dataset used for classification examples.
This is classic titanic dataset used to predict if a passenger will survive or not in titanic ship disaster.
}
\keyword{datasets}
