% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBTrainer.R
\name{NBTrainer}
\alias{NBTrainer}
\title{Naive Bayes Trainer}
\format{\code{\link{R6Class}} object.}
\description{
Trains a naive bayes model. It is built on top high performance naivebayes R package.
}
\section{Usage}{

For usage details see \bold{Methods, Arguments and Examples} sections.
\preformatted{
nbt = NBTrainer$new(prior=NULL, laplace=0, usekernel=FALSE)
nbt$fit(X_train, "target")
prediction <- bst$predict(X_test)
}
}

\section{Methods}{

\describe{
  \item{\code{$new()}}{Initialises an instance of naive bayes model}
  \item{\code{$fit()}}{fits model to an input train data and trains the model.}
  \item{\code{$predict()}}{returns predictions by fitting the trained model on test data.}
}
}

\section{Arguments}{

\describe{
 \item{prior}{for detailed explanation of parameters, check: https://cran.r-project.org/package=naivebayes}
 \item{prior}{numeric vector with prior probabilities. vector with prior probabilities of the classes.
              If unspecified, the class proportions for the training set are used.
              If present, the probabilities should be specified in the order of the factor levels.}
 \item{laplace}{value used for Laplace smoothing. Defaults to 0 (no Laplace smoothing)}
 \item{usekernel}{if TRUE, density is used to estimate the densities of metric predictors}
}
}

\examples{
data(iris)
nb <- NBTrainer$new()
nb$fit(iris, 'Species')
y <- nb$predict(iris)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{NBTrainer$new()}}
\item \href{#method-fit}{\code{NBTrainer$fit()}}
\item \href{#method-predict}{\code{NBTrainer$predict()}}
\item \href{#method-clone}{\code{NBTrainer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NBTrainer$new(prior, laplace, usekernel)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\subsection{Method \code{fit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NBTrainer$fit(X, y)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict"></a>}}
\subsection{Method \code{predict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NBTrainer$predict(X, type = "class")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NBTrainer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
