% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reg_train}
\alias{reg_train}
\title{reg_train}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 1460 rows and 81 columns.
}
\source{
\url{https://www.kaggle.com/c/house-prices-advanced-regression-techniques/data}
}
\usage{
reg_train
}
\description{
Training Dataset used for regression examples. In this data set, we have to predict
the sale price of the houses.
}
\keyword{datasets}
