% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGBoost.R
\docType{data}
\name{XGBTrainer}
\alias{XGBTrainer}
\title{Extreme Gradient Boosting Model}
\format{\code{\link{R6Class}} object.}
\usage{
XGBTrainer
}
\description{
Trains a Extreme Gradient Boosting Model. It is a wrapper around xgboost R package to provide a similar interface to scikit-learn. You simply need to provide a data-frame or data.table to make it work.
}
\section{Usage}{

For usage details see \bold{Methods, Arguments and Examples} sections.
\preformatted{
bst = XGBTrainer$new(n_estimators, max_features, max_depth, min_samples_split, criterion,classification, class_weights, verbose, seed,always_split)
bst$fit(X_train, y_train, valid=NULL)
tfidf$predict(X_test)
}
}

\section{Methods}{

\describe{
  \item{\code{$new()}}{Initialises an instance of xgboost model}
  \item{\code{$fit()}}{fits model to an input train data using given parametes.}
  \item{\code{$cross_val()}}{performs cross validation on train data}
  \item{\code{$predict()}}{returns predictions by fitting the trained model on test data.}
}
}

\section{Arguments}{

\describe{
 \item{params}{the detailed version of these arguments can found in xgboost documentation \url{http://xgboost.readthedocs.io/en/latest/parameter.html}}
 \item{booster}{the trainer type, the values are \code{gbtree(default)}, \code{gblinear}, \code{dart:gbtree}}
 \item{objective}{specify the learning task. Check the link above for all possible values.}
 \item{max_features}{the number of features to consider when looking for the best split
     Possible values are \code{auto},\code{sqrt},\code{log},\code{None}}
 \item{n_thread}{number of parallel threads used to run, default is to run using all threads available}
 \item{silent}{0 means printing running messages, 1 means silent mode}
 \item{n_estimators}{number of trees to grow, default = 100}
 \item{learning_rate}{Step size shrinkage used in update to prevents overfitting. Lower the learning rate, more time it takes in training, value lies between between 0 and 1. Default = 0.3}
 \item{max_depth}{the maximum depth of each tree, default = 6}
 \item{nfold}{set number of folds for cross validation}
 \item{stratified}{if stratified sampling to be done or not, default is TRUE}
 \item{folds}{the list of CV folds' indices - either those passed through the folds parameter or randomly generated.}
 \item{gamma}{Minimum loss reduction required to make a further partition on a leaf node of the tree. The larger gamma is, the more conservative the algorithm will be. Value lies between 0 and infinity, Default = 0}
 \item{min_child_weight}{Minimum sum of instance weight (hessian) needed in a child. If the tree partition step results in a leaf node with the sum of instance weight less than min_child_weight, then the building process will give up further partitioning. In linear regression task, this simply corresponds to minimum number of instances needed to be in each node. The larger min_child_weight is, the more conservative the algorithm will be. Value lies between 0 and infinity. Default = 1}
 \item{subsample}{Subsample ratio of the training instances. Setting it to 0.5 means that XGBoost would randomly sample half of the training data prior to growing trees. and this will prevent overfitting. Subsampling will occur once in every boosting iteration. Value lies between 0 and 1. Default = 1}
 \item{colsample_bytree}{Subsample ratio of columns when constructing each tree. Subsampling will occur once in every boosting iteration. Value lies between 0 and 1. Default = 1}
 \item{lambda}{L2 regularization term on weights. Increasing this value will make model more conservative. Default = 1}
 \item{alpha}{L1 regularization term on weights. Increasing this value will make model more conservative. Default = 0}
 \item{eval_metric}{Evaluation metrics for validation data, a default metric will be assigned according to objective }
 \item{print_every}{print training log after n interations. Default = 50}
 \item{feval}{custom evaluation function}
 \item{early_stopping}{Used to prevent overfitting, stops model training after this number of iterations if there is no improvement seen}
 \item{maximize}{If feval and early_stopping_rounds are set, then this parameter must be set as well. When it is TRUE, it means the larger the evaluation score the better. }
 \item{custom_objective}{custom objective functioon}
 \item{save_period}{when it is non-NULL, model is saved to disk after every save_period rounds, 0 means save at the end.}
 \item{save_name}{the name or path for periodically saved model file.}
 \item{xgb_model}{a previously built model to continue the training from. Could be either an object of class xgb.Booster, or its raw data, or the name of a file with a previously saved model.}
 \item{callbacks}{a list of callback functions to perform various task during boosting. See callbacks. Some of the callbacks are automatically created depending on the parameters' values. User can provide either existing or their own callback methods in order to customize the training process.}
 \item{verbose}{If 0, xgboost will stay silent. If 1, xgboost will print information of performance. If 2, xgboost will print some additional information. Setting verbose > 0 automatically engages the cb.evaluation.log and cb.print.evaluation callback functions.}
 \item{watchlist}{what information should be printed when verbose=1 or verbose=2. Watchlist is used to specify validation set monitoring during training. For example user can specify watchlist=list(validation1=mat1, validation2=mat2) to watch the performance of each round's model on mat1 and mat2}
 \item{num_class}{set number of classes in case of multiclassification problem}
 \item{weight}{a vector indicating the weight for each row of the input.}
 \item{na_missing}{by default is set to NA, which means that NA values should be considered as 'missing' by the algorithm. Sometimes, 0 or other extreme value might be used to represent missing values. This parameter is only used when input is a dense matrix.}
}
}

\examples{
library(data.table)
df <- copy(iris)

# convert characters/factors to numeric
df$Species <- as.numeric(as.factor(df$Species))-1

# initialise model
xgb <- XGBTrainer$new(objective = 'multi:softmax',
                      maximize = FALSE,
                      eval_metric = 'merror',
                      num_class=3,
                      n_estimators = 500)
xgb$fit(df, 'Species')

# do cross validation to find optimal value for n_estimators
xgb$cross_val(X = df, y = 'Species',nfolds = 3, stratified = TRUE, early_stopping = 20)
best_iter <- xgb$cv_model$best_iteration
xgb$show_importance()

# make predictions
preds <- xgb$predict(as.matrix(iris[,1:4]))
preds
}
\keyword{datasets}
